feat: simplify event date handling
