+++ docs/docs/auto-docs/components/UserPortal/ChatRoom/ChatRoom.fixtures/variables/CHAT_BY_ID_ERROR_MOCK.md
@@ -1,0 +1,47 @@
+[Admin Docs](/)
+
+***
+
+# Variable: CHAT\_BY\_ID\_ERROR\_MOCK
+
+> `const` **CHAT\_BY\_ID\_ERROR\_MOCK**: `object`
+
+Defined in: [src/components/UserPortal/ChatRoom/ChatRoom.fixtures.ts:528](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/ChatRoom/ChatRoom.fixtures.ts#L528)
+
+## Type Declaration
+
+### error
+
+> **error**: `Error`
+
+### request
+
+> **request**: `object`
+
+#### request.query
+
+> **query**: `DocumentNode` = `CHAT_BY_ID`
+
+#### request.variables
+
+> **variables**: `object`
+
+#### request.variables.beforeMessages
+
+> **beforeMessages**: `any` = `null`
+
+#### request.variables.first
+
+> **first**: `number` = `15`
+
+#### request.variables.input
+
+> **input**: `object`
+
+#### request.variables.input.id
+
+> **id**: `string` = `'chat123'`
+
+#### request.variables.lastMessages
+
+> **lastMessages**: `number` = `15`
