+++ docs/docs/auto-docs/components/UserPortal/ChatRoom/ChatRoom.fixtures/variables/FIXED_UTC.md
@@ -1,0 +1,9 @@
+[Admin Docs](/)
+
+***
+
+# Variable: FIXED\_UTC
+
+> `const` **FIXED\_UTC**: `string`
+
+Defined in: [src/components/UserPortal/ChatRoom/ChatRoom.fixtures.ts:10](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/ChatRoom/ChatRoom.fixtures.ts#L10)
