+++ docs/docs/auto-docs/components/UserPortal/ChatRoom/ChatRoom.fixtures/variables/FIXED_UTC_MINUS_ONE_DAY.md
@@ -1,0 +1,9 @@
+[Admin Docs](/)
+
+***
+
+# Variable: FIXED\_UTC\_MINUS\_ONE\_DAY
+
+> `const` **FIXED\_UTC\_MINUS\_ONE\_DAY**: `string`
+
+Defined in: [src/components/UserPortal/ChatRoom/ChatRoom.fixtures.ts:11](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/ChatRoom/ChatRoom.fixtures.ts#L11)
