+++ docs/docs/auto-docs/components/UserPortal/ChatRoom/ChatRoom.fixtures/variables/UNREAD_CHATS_MOCK.md
@@ -1,0 +1,35 @@
+[Admin Docs](/)
+
+***
+
+# Variable: UNREAD\_CHATS\_MOCK
+
+> `const` **UNREAD\_CHATS\_MOCK**: `object`
+
+Defined in: [src/components/UserPortal/ChatRoom/ChatRoom.fixtures.ts:188](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/ChatRoom/ChatRoom.fixtures.ts#L188)
+
+## Type Declaration
+
+### request
+
+> **request**: `object`
+
+#### request.query
+
+> **query**: `DocumentNode` = `UNREAD_CHATS`
+
+#### request.variables
+
+> **variables**: `object` = `{}`
+
+### result
+
+> **result**: `object`
+
+#### result.data
+
+> **data**: `object`
+
+#### result.data.unreadChats
+
+> **unreadChats**: `any`[] = `[]`
