+++ docs/docs/auto-docs/types/Event/interface/type-aliases/InterfaceAttendanceStatisticsModalProps.md
@@ -6,4 +6,4 @@
 
 > **InterfaceAttendanceStatisticsModalProps** = [`IAttendanceStatisticsModalProps`](../interfaces/IAttendanceStatisticsModalProps.md)
 
-Defined in: [src/types/Event/interface.ts:308](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L308)
+Defined in: [src/types/Event/interface.ts:305](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L305)
