+++ docs/docs/auto-docs/types/Event/interface/type-aliases/InterfaceCalendarProps.md
@@ -6,4 +6,4 @@
 
 > **InterfaceCalendarProps** = [`ICalendarProps`](../interfaces/ICalendarProps.md)
 
-Defined in: [src/types/Event/interface.ts:302](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L302)
+Defined in: [src/types/Event/interface.ts:299](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L299)
