+++ docs/docs/auto-docs/types/Event/interface/type-aliases/InterfaceDeleteEventModalProps.md
@@ -6,4 +6,4 @@
 
 > **InterfaceDeleteEventModalProps** = [`IDeleteEventModalProps`](../interfaces/IDeleteEventModalProps.md)
 
-Defined in: [src/types/Event/interface.ts:304](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L304)
+Defined in: [src/types/Event/interface.ts:301](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L301)
