+++ docs/docs/auto-docs/types/Event/interface/type-aliases/InterfaceEvent.md
@@ -6,4 +6,4 @@
 
 > **InterfaceEvent** = [`IEvent`](../interfaces/IEvent.md)
 
-Defined in: [src/types/Event/interface.ts:299](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L299)
+Defined in: [src/types/Event/interface.ts:296](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L296)
