+++ docs/docs/auto-docs/types/Event/interface/type-aliases/InterfaceEventEdge.md
@@ -6,4 +6,4 @@
 
 > **InterfaceEventEdge** = [`IEventEdge`](../interfaces/IEventEdge.md)
 
-Defined in: [src/types/Event/interface.ts:306](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L306)
+Defined in: [src/types/Event/interface.ts:303](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L303)
