+++ docs/docs/auto-docs/types/Event/interface/type-aliases/InterfaceEventHeaderProps.md
@@ -6,4 +6,4 @@
 
 > **InterfaceEventHeaderProps** = [`IEventHeaderProps`](../interfaces/IEventHeaderProps.md)
 
-Defined in: [src/types/Event/interface.ts:303](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L303)
+Defined in: [src/types/Event/interface.ts:300](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L300)
