+++ docs/docs/auto-docs/types/Event/interface/type-aliases/InterfaceMember.md
@@ -6,4 +6,4 @@
 
 > **InterfaceMember** = [`IMember`](../interfaces/IMember.md)
 
-Defined in: [src/types/Event/interface.ts:298](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L298)
+Defined in: [src/types/Event/interface.ts:295](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L295)
