+++ docs/docs/auto-docs/types/Event/interface/type-aliases/InterfaceStatsModal.md
@@ -6,4 +6,4 @@
 
 > **InterfaceStatsModal** = [`IStatsModal`](../interfaces/IStatsModal.md)
 
-Defined in: [src/types/Event/interface.ts:301](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L301)
+Defined in: [src/types/Event/interface.ts:298](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L298)
