+++ docs/docs/auto-docs/types/Event/interface/type-aliases/InterfaceUpdateEventModalProps.md
@@ -6,4 +6,4 @@
 
 > **InterfaceUpdateEventModalProps** = [`IUpdateEventModalProps`](../interfaces/IUpdateEventModalProps.md)
 
-Defined in: [src/types/Event/interface.ts:307](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L307)
+Defined in: [src/types/Event/interface.ts:304](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/Event/interface.ts#L304)
