+++ src/screens/AdminPortal/OrganizationEvents/CreateEventModal.tsx
@@ -111,24 +111,17 @@ const CreateEventModal: React.FC<ICreateEventModalProps> = ({
       // Build input object with shared typed interface
       const input: ICreateEventInput = {
         name: payload.name,
+        startAt: payload.startAtISO,
+        endAt: payload.endAtISO,
         organizationId: currentUrl,
         allDay: payload.allDay,
         isPublic: payload.isPublic,
         isRegisterable: payload.isRegisterable,
-        isInviteOnly: payload.isInviteOnly,
 
         ...(payload.description && { description: payload.description }),
         ...(payload.location && { location: payload.location }),
         ...(recurrenceInput && { recurrence: recurrenceInput }),
       };
-
-      if (payload.allDay) {
-        input.startDate = payload.startAtISO;
-        input.endDate = payload.endAtISO;
-      } else {
-        input.startAt = payload.startAtISO;
-        input.endAt = payload.endAtISO;
-      }
 
       const { data: createEventData } = await create({
         variables: { input },
